fix to avoid integer maths

+++ src/Analyses/STAR_2009_UE_HELEN.cc
@@ -118,7 +118,7 @@ namespace Rivet {
         //        for the STAR tracking efficiency. Once we have the
         //        final numbers (corrected to hadron level), we need
         //        to remove this!!!!
-        if (1.0*rand()/RAND_MAX > 0.87834-exp(-1.48994-0.788432*pT)) {
+        if (1.0*rand()/static_cast<double>(RAND_MAX) > 0.87834-exp(-1.48994-0.788432*pT)) {
           continue;
         }
         // -------- end of efficiency hack -------
